+++ CHANGELOG.md
@@ -28,6 +28,7 @@
 * Prodigal Memory (thanks LankSSBM)
 * Silver Strike (thanks LankSSBM)
 * Ticking Curse (thanks LankSSBM)
+* Sacrifice (thanks LankSSBM)
 * Shrapnel Bloom (thanks LankSSBM)
 * Assertion (thanks LankSSBM)
 * Investment (thanks LankSSBM)